# [libc++] Refactor num_get optimization to not be ABI breaking